alert when trying to delete a user or trying to log out

+++ components/Settings.tsx
@@ -1,5 +1,6 @@
 import MyDrawer from "./Menu";
-import { Text, View, Button, Popover } from 'native-base';
+import { Text, View, Button } from 'native-base';
+import { Alert } from "react-native";
 import React from 'react';
 import { styles } from './StyleSheet'
 import AsyncStorage from "@react-native-async-storage/async-storage";
@@ -19,9 +20,25 @@ interface User {
 
 export default function Settings(props: any) {
   const kirjauduUlos = async () => {
-    await AsyncStorage.removeItem('token');
-    props.setLoggedIn(false);
-  }
+    Alert.alert(
+      'Haluatko varmasti kirjautua ulos?',
+      '',
+      [
+        {
+          text: 'Kyllä',
+          onPress: async () => {
+            await AsyncStorage.removeItem('token');
+            props.setLoggedIn(false);
+          },
+        },
+        {
+          text: 'Ei',
+          style: 'cancel',
+        },
+      ],
+    );
+  };
+
 
   const poistaTunnus = async () => {
     try {
@@ -30,42 +47,57 @@ export default function Settings(props: any) {
         console.log('No token found');
         return;
       }
-
+  
       const response = await fetch(userUrl, {
         method: 'GET',
         headers: {
           Authorization: `Bearer ${token}`,
         },
       });
-
+  
       if (!response.ok) {
         console.log('Error getting current user');
         return;
       }
-
+  
       const currentUser: User = await response.json();
       console.log(currentUser);
-
-      const deleteUrl = `https://dev-discgolf.herokuapp.com/users/${currentUser.userId}`;
-      const deleteResponse = await fetch(deleteUrl, {
-        method: 'DELETE',
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      });
-
-      if (deleteResponse.ok) {
-        await AsyncStorage.removeItem('token');
-        props.setLoggedIn(false);
-        console.log('User removed successfully');
-      } else {
-        console.log('Error removing user');
-      }
+  
+      Alert.alert(
+        'Haluatko varmasti poistaa tämän tunnuksen?',
+        'Tämä toiminto poistaa käyttäjän sekä kaiken sen datan.',
+        [
+          {
+            text: 'Kyllä',
+            onPress: async () => {
+              const deleteUrl = `https://dev-discgolf.herokuapp.com/users/${currentUser.userId}`;
+              const deleteResponse = await fetch(deleteUrl, {
+                method: 'DELETE',
+                headers: {
+                  Authorization: `Bearer ${token}`,
+                },
+              });
+  
+              if (deleteResponse.ok) {
+                await AsyncStorage.removeItem('token');
+                props.setLoggedIn(false);
+                console.log('User removed successfully');
+              } else {
+                console.log('Error removing user');
+              }
+            }
+          },
+          {
+            text: 'Ei',
+            style: 'cancel'
+          },
+        ]
+      );
     } catch (error) {
       console.error(error);
     }
   };
-
+  
   return (
     <View style={styles.view}>
       <Button _pressed={{ opacity: 0.5 }} style={styles.button}>Nollaa tilastot</Button>
@@ -75,4 +107,4 @@ export default function Settings(props: any) {
       <Button _pressed={{ opacity: 0.5 }} style={styles.button} onPress={kirjauduUlos}>Kirjaudu ulos</Button>
     </View>
   );
-}
+  }
